test: platform-gate StartJob tests and cover the GC live-job invariant (#33)
- Gate the mcptools tests that drive StartJob (Linux-only job
  supervision) behind the _linux filename suffix: rename
  run_sync_test.go to run_sync_linux_test.go (it also holds the
  connect/newTestManager helpers), and move TestAgyRunAndStatusOverMCP
  out of tools_test.go into tools_linux_test.go. The cross-platform
  tests (toStartRequest validation in tools_test.go, HTTP serving in
  serve_http_test.go) stay ungated, so `go test ./...` is green on
  macOS and Windows instead of failing on the unsupported-platform
  StartJob path.
- Add TestGarbageCollectKeepsExpiredButAliveJob: GC's core safety
  invariant (an expired job whose supervisor is still alive must be
  kept) was untested because every other GC test uses dead metas. This
  one spawns a real live process so a regression dropping the liveness
  check is caught.

Test-only. The remaining #33 item (the supervisor SIGKILL-escalation
test, which needs killGrace made injectable plus a SIGTERM-ignoring
fake agy) is left as a follow-up.

Refs #33
