+++ internal/mcptools/run_sync_linux_test.go
